LayoutTests/css3/color-filters/color-filter-exposed-if-disabled-expected.txt

Tests that -apple-color-filter is not exposed when the feature is disabled

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL '-apple-color-filter' in document.documentElement.style should be false. Was true.
FAIL '-apple-color-filter' in getComputedStyle(document.documentElement) should be false. Was true.
PASS CSS.supports('-apple-color-filter: contrast(1)') is false
PASS CSS.supports('-apple-color-filter: inherit') is false
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
